+++ pyqgiswps/ogc/ogc.py
@@ -41,8 +41,13 @@ OGCTYPE = {
 
 OGCUNIT = {
     'degree': 'urn:ogc:def:uom:OGC:1.0:degree',
+    'degrees': 'urn:ogc:def:uom:OGC:1.0:degree',
+    'meter': 'urn:ogc:def:uom:OGC:1.0:metre',
     'metre': 'urn:ogc:def:uom:OGC:1.0:metre',
-    'unity': 'urn:ogc:def:uom:OGC:1.0:unity'
+    'meteres': 'urn:ogc:def:uom:OGC:1.0:metre',
+    'meters': 'urn:ogc:def:uom:OGC:1.0:metre',
+    'unity': 'urn:ogc:def:uom:OGC:1.0:unity',
+    'feet': 'urn:ogc:def:uom:OGC:1.0:feet'
 }